TDI-41443 : migrating bigquery to TCK
Fixed NPE on Table data set

+++ bigquery/src/main/java/org/talend/components/bigquery/input/BigQueryTableExtractInput.java
@@ -131,7 +131,9 @@ public class BigQueryTableExtractInput implements Serializable {
 
     @PreDestroy
     public void release() {
-        storageService.deleteBlob(storage, bucket, gsBlob);
+        if (gsBlob != null) {
+            storageService.deleteBlob(storage, bucket, gsBlob);
+        }
     }
 
 }
